app_backend/requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
flask
flask-cors